assign role to user

+++ admin/src/main/java/com/admin/user/vo/UserVO.java
@@ -2,11 +2,13 @@ package com.admin.user.vo;
 
 import java.io.Serializable;
 import java.util.Date;
+import java.util.List;
 
 import com.admin.helper.vo.BaseVO;
+import com.admin.role.vo.RoleVO;
 
 public class UserVO extends BaseVO implements Serializable {
-    private Integer userId;
+    private Long userId;
 
     private String userName;
 
@@ -26,11 +28,11 @@ public class UserVO extends BaseVO implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
